refactor(cli): unify simulate and local-testbed renderers
Generify Terminal over a new ConfigRender trait so both commands flow
through the same print_config/print_status/print_results/print_summary
pipeline. Move all presentation traits into terminal/render.rs (renamed
from result.rs to reflect the broader scope) and delete status.rs (its
body is now MetricsSnapshotsRender on [MetricsSnapshot]). Spinner is
long-lived with start/stop semantics, hidden via ProgressBar::hidden()
on non-TTY stderr.

Closes #64
